% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FiSh.R
\docType{package}
\name{FiSh-package}
\alias{FiSh}
\alias{FiSh-package}
\title{FiSh: Fisher-Shannon method}
\description{
Proposes non-parametric estimates of the Fisher Information 
          Measure and the Shannon Entropy Power. A Python version of 
          this work is available on GitHub and PyPi (FiShPy).
}
\details{
This R code was developed and used for the following paper:
F. Guignard, M. Laib, F. Amato, M. Kanevski, Advanced analysis of temporal data using Fisher-Shannon
information : theoretical development and application to geoscience, In prep.
}
\references{
S. J. Sheather and M. C. Jones (1991). A reliable data-based bandwidth selection method for kernel density
estimation. Journal of the Royal Statistical Society, Series B, 53, 683 - 690.

M. P. Wand and M. C. Jones (1995). Kernel Smoothing. Chapman and Hall, London.

C. Vignat, J.F Bercher (2003). Analysis of signals in the Fisher–Shannon information plane, Physics Letters A,
312, 190, 27 – 33.

F. Guignard, M. Laib, F. Amato, M. Kanevski, Advanced analysis of temporal data using Fisher-Shannon
information : theoretical development and application to geoscience, In prep.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://FiShInfo.github.io/}
}

}
\author{
Fabian Guignard \email{fabian.guignard@bluemail.ch} and

Mohamed Laib \email{Mohamed.Laib@gmail.ch}

Maintainer: Mohamed Laib \email{laib.med@gmail.com}
}
