% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FiSh.R
\docType{package}
\name{FiSh-package}
\alias{FiSh}
\alias{FiSh-package}
\title{FiSh: Fisher-Shannon Method}
\description{
Proposes non-parametric estimates of the Fisher information measure and the 
              Shannon entropy power. More theoretical and implementation details can be found
              in Guignard et al. <doi:10.3389/feart.2020.00255>. A 'python' version of this
              work is available on 'github' and 'PyPi' ('FiShPy').
}
\details{
If this R code is used for academic research, please cite the following paper where it was developed:

F. Guignard, M. Laib, F. Amato, M. Kanevski, Advanced analysis of temporal data using Fisher-Shannon information:
theoretical development and application in geosciences, 2020,
\doi{10.3389/feart.2020.00255}{Frontiers in Earth Science, 8:255}.
}
\references{
S. J. Sheather and M. C. Jones (1991). A reliable data-based bandwidth selection method for kernel density
estimation. Journal of the Royal Statistical Society, Series B, 53, 683 - 690.

M. P. Wand and M. C. Jones (1995). Kernel Smoothing. Chapman and Hall, London.

C. Vignat, J.F Bercher (2003). Analysis of signals in the Fisher–Shannon information plane, Physics Letters A,
312, 190, 27 – 33.

F. Guignard, M. Laib, F. Amato, M. Kanevski, Advanced analysis of temporal data using Fisher-Shannon information:
theoretical development and application in geosciences, 2020,
\doi{10.3389/feart.2020.00255}{Frontiers in Earth Science, 8:255}.
}
\author{
Fabian Guignard \email{fabian.guignard@protonmail.ch} and

Mohamed Laib \email{laib.med@gmail.com}

Maintainer: Mohamed Laib \email{laib.med@gmail.com}
}
